requirements.txt
streamlit
pandas
numpy
xgboost
scikit-learn
pydrive2
matplotlib
shap
pdfkit
plotly
python-dotenv